Change memory attribute
Non cacheable memory was used before. It leads to an exception
if an exclusive load/store operation was used. It also disallows
to use a spin lock because the spin lock is based on an exclusive
load/store operation.

For spin lock usage the memory attributes have been changed.

Issue: #MCOM03SW-198
Change-Id: Ibb026476a0f1c628c3fa600fb98fb333bd2e7887