mcom03: Add support for TSP as BL32
TF-A provides TSP as an example of how a Trusted OS is supported
on a production system using the Runtime Services Framework.
On such a system, the Test BL32 image and service are replaced by
the Trusted OS and its dispatcher service.

Also see: https://trustedfirmware-a.readthedocs.io/en/latest/design/firmware-design.html#secure-el1-payloads-and-dispatchers

Issue: #MCOM03SW-1483
Change-Id: I4bd4e2a9f5dff9dd6f0996340427ffc054077598